+++ example/book_store/test/helpers.dart
@@ -1,4 +1,3 @@
-import 'dart:collection';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';